% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapUtils.R
\name{addBarChart}
\alias{addBarChart}
\alias{updateBarChart}
\title{Add or update polar charts}
\usage{
addBarChart(map, lng, lat, data, size = 30, opacity = 1, minValue = NULL,
  maxValue = NULL, colors = NULL, popup = NULL, layerId = NULL)

addBarChart(map, lng, lat, data, size = 30, opacity = 1, minValue = NULL,
  maxValue = NULL, colors = NULL, popup = NULL, layerId = NULL)

updateBarChart(map, layerId, data = NULL, size = NULL, opacity = NULL,
  minValue = NULL, maxValue = NULL, colors = NULL, popup = NULL)
}
\arguments{
\item{map}{A map object created with \code{\link[leaflet]{leaflet()}}}

\item{lng}{longitude of the polar charts}

\item{lat}{lattitude of the polar charts}

\item{data}{numeric matrix. Number of rows must equal the number of polar charts. The
number of columns is equal to the number of sectors in each polar chart.}

\item{opacity}{Opacity of a polar chart}

\item{maxValue}{Either a single numeric value, or a vector with length equal to the number
of columns of data. In the first case (one value), all variable will share
the same scale while in the second one each variable will have its own scale.
If it is \code{NULL}, the maximal value of each column is used (so scales
are different for each column). If it equals to 0, the global maximum of
the matrix is used and all variables share the same scale.}

\item{colors}{Vector of colors for the sectors of the polar chart}

\item{popup}{a character vector of the HTML content for the popups. They are displayed 
when user clicks on a segment.}

\item{layerId}{Layer id.}

\item{radius}{maximal radius of a polar chart}

\item{scale}{Should the values be represented by the radius or the area ? Radius by 
default. This creates a high distorsion of the values, but permits to 
clearly see differences between different zones.}

\item{lng}{longitude of the polar charts}

\item{lat}{lattitude of the polar charts}

\item{data}{numeric matrix. Number of rows must equal the number of polar charts. The
number of columns is equal to the number of sectors in each polar chart.}

\item{radius}{maximal radius of a polar chart}

\item{opacity}{Opacity of a polar chart}

\item{scale}{Should the values be represented by the radius or the area ? Radius by 
default. This creates a high distorsion of the values, but permits to 
clearly see differences between different zones.}

\item{maxValue}{Either a single numeric value, or a vector with length equal to the number
of columns of data. In the first case (one value), all variable will share
the same scale while in the second one each variable will have its own scale.
If it is \code{NULL}, the maximal value of each column is used (so scales
are different for each column). If it equals to 0, the global maximum of
the matrix is used and all variables share the same scale.}

\item{colors}{Vector of colors for the sectors of the polar chart}
}
\value{
The modified map object.

The modified map object.
}
\description{
These functions add and update polar area charts on a leaflet map.

These functions add and update polar area charts on a leaflet map.
}
\examples{

require(leaflet)
mydata <- rbind(1:3, 4:6)

# Polar chart with distinct scales for each column of mydata. For the second
# polar chart the radius will be equal to 20 (maximal radius) for each sector.
# For the first one radius will be 1/4 * 20, 2/5 * 20 and 3/6 * 20

leaflet() \%>\% addPolarCharts(c(0,30), c(0, 0), mydata) 

# Common scale for the three columns

leaflet() \%>\% addPolarCharts(c(0,30), c(0, 0), mydata, maxValue = 0)

# Explicitely provide the maximal value. This is useful for comparison to have
# the same scale on different maps.

leaflet() \%>\% addPolarChart(c(0,30), c(0, 0), mydata, maxValue = 9)


require(leaflet)
mydata <- rbind(1:3, 4:6)

# Polar chart with distinct scales for each column of mydata. For the second
# polar chart the radius will be equal to 20 (maximal radius) for each sector.
# For the first one radius will be 1/4 * 20, 2/5 * 20 and 3/6 * 20

leaflet() \%>\% addPolarCharts(c(0,30), c(0, 0), mydata) 

# Common scale for the three columns

leaflet() \%>\% addPolarCharts(c(0,30), c(0, 0), mydata, maxValue = 0)

# Explicitely provide the maximal value. This is useful for comparison to have
# the same scale on different maps.

leaflet() \%>\% addPolarChart(c(0,30), c(0, 0), mydata, maxValue = 9)

}

