% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletPlugins.R
\name{addBarCharts}
\alias{addBarCharts}
\alias{updateBarCharts}
\title{Add or update bar charts}
\usage{
addBarCharts(map, lng, lat, data, size = 30, opacity = 1, minValue = NULL,
  maxValue = NULL, colors = NULL, popup = NULL, layerId = NULL)

updateBarCharts(map, layerId, data = NULL, size = NULL, opacity = NULL,
  minValue = NULL, maxValue = NULL, colors = NULL, popup = NULL)
}
\arguments{
\item{map}{A map object created with \code{\link[leaflet]{leaflet()}}}

\item{lng}{longitude of the bar charts}

\item{lat}{lattitude of the bar charts}

\item{data}{numeric matrix. Number of rows must equal the number of bar charts. The
number of columns is equal to the number of bars in each bar chart.}

\item{size}{maximal width and height of the bar charts}

\item{opacity}{Opacity of a bar chart}

\item{maxValue}{Either a single numeric value, or a vector with length equal to the number
of columns of data. In the first case (one value), all variable will share
the same scale while in the second one each variable will have its own scale.
If it is \code{NULL}, the maximal value of each column is used (so scales
are different for each column). If it equals to 0, the global maximum of
the matrix is used and all variables share the same scale.}

\item{colors}{Vector of colors for each bar of the bar charts}

\item{popup}{a character vector of the HTML content for the popups. They are displayed 
when user clicks on a segment.}

\item{layerId}{Layer id.}
}
\value{
The modified map object.
}
\description{
These functions add and update bar charts on a leaflet map.
}
\examples{

require(leaflet)
mydata <- rbind(1:3, 4:6)

# Bar chart with distinct scales for each column of mydata. For the second
# bar chart the height will be equal to 20 (maximal size) for each bar.
# For the first one height will be 1/4 * 20, 2/5 * 20 and 3/6 * 20

leaflet() \%>\% addBarChart(c(0,30), c(0, 0), mydata) 

# Common scale for the three columns

leaflet() \%>\% addBarCharts(c(0,30), c(0, 0), mydata, maxValue = 0)

# Explicitely provide the maximal value. This is useful for comparison to have
# the same scale on different maps.

leaflet() \%>\% addBarCharts(c(0,30), c(0, 0), mydata, maxValue = 50)

}

