% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchangesStack.R
\name{exchangesStack}
\alias{exchangesStack}
\title{Plot the exchanges of an area}
\usage{
exchangesStack(x, area = NULL, dateRange = NULL, colors = NULL,
  main = NULL, ylab = NULL, unit = c("MWh", "GWh", "TWh"),
  interactive = base::interactive(), legend = TRUE,
  legendId = sample(1e+09, 1), legendItemsPerRow = 5, width = NULL,
  height = NULL)
}
\arguments{
\item{x}{data.table of class \code{antaresDataTable} created with function
\code{\link[antaresRead]{readAntares}}. It must contain link data.}

\item{area}{Name of a single area. The flows from/to this area will be drawn by the
function.}

\item{dateRange}{A vector of two dates. Only data points between these two dates are 
displayed. If NULL, then all data is displayed.}

\item{colors}{Vector of colors with same length as parameter \code{variables}. If 
\code{variables} is an alias, then this argument should be \code{NULL} in 
order to use default colors.}

\item{main}{Title of the graph.}

\item{unit}{Unit used in the graph. Possible values are "MWh", "GWh" or "TWh".}

\item{interactive}{LogicalValue. If \code{TRUE}, then a shiny gadget is launched that lets
the user interactively choose the areas or districts to display.}

\item{legend}{Logical value indicating if a legend should be drawn. This argument is 
usefull when one wants to create a shared legend with
\code{\link{productionStackLegend}}}

\item{legendId}{Id of the legend linked to the graph. This argument is 
usefull when one wants to create a shared legend with
\code{\link{productionStackLegend}}}

\item{legendItemsPerRow}{Number of elements to put in each row of the legend.}

\item{width}{Width of the graph expressed in pixels or in percentage of 
the parent element. For instance "500px" and "100\%" are valid values.}

\item{height}{Height of the graph expressed in pixels or in percentage of 
the parent element. For instance "500px" and "100\%" are valid values.}
}
\value{
A htmlwidget of class \code{dygraph}. It can be modified with functions from
package \code{dygraphs}.
}
\description{
This function draws a stack representing the evolution of the exchanges of
an area with its neighbours.
}
\examples{
\dontrun{
mydata <- readAntares(links = "all", timeStep = "daily")
exchangeStack(mydata)

}

}

