% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limitSizeGraph.R
\name{limitSizeGraph}
\alias{limitSizeGraph}
\title{Use to change limit size of graph}
\usage{
limitSizeGraph(size)
}
\arguments{
\item{size}{\code{numeric} size autorize of graphs in modules (default 1)}
}
\description{
Use to change limit size of graph
}
\examples{
\dontrun{
limitSizeGraph(10)
}

}
