% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layout.R
\name{mapLayout_no_interactive}
\alias{mapLayout_no_interactive}
\title{Build a `mapLayout` object with gesojson file}
\usage{
mapLayout_no_interactive(path_geojson_file, opts = simOptions())
}
\arguments{
\item{path_geojson_file}{`character` path of geojson file}

\item{opts}{list of simulation parameters returned by 
the function \code{\link[antaresRead]{setSimulationPath}}}
}
\value{
Object of class "mapLayout"
}
\description{
This function creates a 'mapLayout' object from a study and an external 
'geojson' file.  
 
This function should be used only once per study.  

The result should then be saved in an external file and be reused.
}
\note{
The 'geojson' file must contain zones compatible with the study.
}
\examples{
\dontrun{
# set informations to your study ("input" mode is enough)
setSimulationPath(path = "path/my_study", simulation = "input")

path_geojson <- "path/my_geosjonfile.geojson"

mapLayout_no_interactive(path_geojson_file = path_geojson)

}
}
