% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rpart.R
\name{modRpart}
\alias{modRpart}
\title{Make rpart from antares data}
\usage{
modRpart(data)
}
\arguments{
\item{data}{an antaresData after use of \code{\link[antaresProcessing]{mergeAllAntaresData}}}
}
\description{
Make rpart from antares data
}
\examples{
\dontrun{
setSimulationPath("Mystud", 1)
mydata <- readAntares(areas = "all", select = "OIL")
mydata <- mergeAllAntaresData(mydata)
modRpart(mydata)
}

}
