% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_XY.R
\name{modXY}
\alias{modXY}
\title{Make X-Y bockey plot, interactive version}
\usage{
modXY(x, xyCompare = c("union", "intersect"))
}
\arguments{
\item{x}{optsH5 or list of optsH5}

\item{xyCompare}{Use when you compare studies, can be "union" or "intersect". If union, all
of mcYears in one of studies will be selectable. If intersect, only mcYears in all
studies will be selectable.}
}
\description{
Make X-Y bockey plot, interactive version
}
\examples{
\dontrun{
opts <- setSimulationPath("h5File")
modXY(opts)
modXY(list(opts, opts))

}

}
