% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_antaresDataTable.R
\name{plot.antaresDataTable}
\alias{plot.antaresDataTable}
\title{plot time series contained in an antaresDataTable}
\usage{
\method{plot}{antaresDataTable}(x, variable, main, ylab, ...)
}
\arguments{
\item{x}{Object of class \code{antaresDataTable}.}

\item{variable}{Name of the variable to plot. If this argument is missing, then the 
function starts a shiny gadget that let the user choose the variable to
represent. When the user clicks on the "Done" button", the graphic is
returned by the function.}

\item{main}{Title of the generated graphic}

\item{ylab}{label of the "y" axis}

\item{...}{currently unused}
}
\value{
The function returns an object of class "htmlwidget". It is generated by
package \code{highcharter} if time step is annual or by \code{dygraphs} for 
any other time step.It can be directly displayed in the viewer or be stored
in a variable for later use.
}
\description{
This function generates an interactive plot of an antares Time series.
}
\details{
If the input data contains several Monte-Carlo scenarios, the function will
display the evolution of the average value. Moreover it will represent a
95% confidence interval.

If the input data has a annual time step, the function creates a barplot
instead of a line chart.
}
\examples{
\dontrun{
setSimulationPath()
mydata <- readAntares("all", timeStep = "monthly")
plot(mydata)
plot(mydata, "LOAD")

# Plot only a few areas
plot(mydata[area \%in\% c("area1", "area2", "area3")])

# If data contains detailed results, then the function adds a confidence
# interval
dataDetailed <- readAntares("all", timeStep = "monthly", synthesis = FALSE)
plot(dataDetailed)

# If the time step is annual, the function creates a barplot instead of a
# linechart
dataAnnual <- readAntares("all", timeStep = "Annual")
plot(dataAnnual)

}

}

