% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapLayout.R
\name{plot.mapLayout}
\alias{plot.mapLayout}
\title{Plot method for map layout}
\usage{
\method{plot}{mapLayout}(x, colAreas = x$coords$color, sizeAreas = 10,
  colLinks = "#CCCCCC", sizeLinks = 3, dirLinks = 0, ...)
}
\arguments{
\item{x}{Object created with function \code{\link{mapLayout}}}

\item{colAreas}{Vector of colors for areas. By default, the colors used in the Antares
software are used.}

\item{sizeAreas}{Either a single numeric value indicating the radius of the areas on the 
map or a vector with as many values as the number of areas of the study or
a matrix with as many rows as the number of areas. In the last case, areas 
are represented by polar area charts and each column is represented by one
section of the polar charts.}

\item{colLinks}{Vector of colors for links.}

\item{sizeLinks}{Line width of the links, in pixels.}

\item{dirLinks}{Single value or vector indicating the direction of the link. Possible values
are 0, -1 and 1. If it equals 0, then links are repsented by a simple line. 
If it is equal to 1 or -1 it is represented by a line with an arrow pointing
respectively the destination and the origin of the link.}

\item{...}{Currently unused.}
}
\value{
The function generates an \code{htmlwidget} of class \code{leaflet}. It can
  be stored in a variable and modified with package 
  \code{\link[leaflet]{leaflet}}
}
\description{
This method can be used to visualize the network of an antares study.
It generates an interactive map with a visual representaiton of a
map layout created with function \code{\link{mapLayout}}.
}
\examples{
\dontrun{
# Read the coordinates of the areas in the Antares interface, then convert it
# in a map layout.
layout <- readLayout()
ml <- mapLayout(layout)

# Save the result for future use
save(ml, file = "ml.rda")

# Plot the network on an interactive map
plot(ml)

# change style
plot(ml, colAreas = gray(0.5), colLinks = "orange")

# Use polar area charts to represent multiple values for each area.
nareas <- nrow(ml$coords)
fakeData <- matrix(runif(nareas * 3), ncol = 3)
plot(ml, sizeAreas = fakeData)

# Store the result in a variable to change it with functions from leaflet 
# package
library(leaflet)

center <- c(mean(ml$coords$x), mean(ml$coords$y))

p <- plot(ml)
p \%>\% 
  addCircleMarker(center[1], center[2], color = "red", 
                  popup = "I'm the center !")
}

}

