% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap.R
\name{plotMap}
\alias{plotMap}
\title{Display results on a map}
\usage{
plotMap(x, mapLayout, areaVar = "none", linkVar = "none",
  timeId = min(x$areas$timeId), interactive = base::interactive(),
  mp = NULL)
}
\arguments{
\item{x}{Object of class \code{antaresDataList} created with 
\code{\link[antaresRead]{readAntares}}}

\item{mapLayout}{Object created with function \code{\link{mapLayout}}}

\item{areaVar}{area variable to display on the map}

\item{linkVar}{link variable to display on the map}

\item{timeId}{timeId}

\item{interactive}{Should the function start a shiny gadget that lets the user modify the
parameters of the plot ?}
}
\description{
Display results on a map
}

