% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layout.R
\name{plotMapLayout}
\alias{plotMapLayout}
\title{Visualize mapLayout output.}
\usage{
plotMapLayout(mapLayout)
}
\arguments{
\item{mapLayout}{object returned by function \code{\link{mapLayout}}}
}
\description{
Visualize mapLayout output.
}
\examples{

\dontrun{
# Read the coordinates of the areas in the Antares interface, then convert it
# in a map layout.
layout <- readLayout()
ml <- mapLayout(layout)

# visualize mapLayout
plotMapLayout(ml)

}

}
\seealso{
\code{\link{mapLayout}}
}
