% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMapOptions.R
\name{plotMapOptions}
\alias{plotMapOptions}
\title{Graphical options for plotMap}
\usage{
plotMapOptions(colArea = "#CCCCCC", sizeArea = 15, maxSizeArea = 15,
  colLink = "#CCCCCC", sizeLink = 3, maxSizeLink = 10)
}
\arguments{
\item{colArea}{default color of areas.}

\item{sizeArea}{default size of areas.}

\item{maxSizeArea}{maximal size of an area when it represents the value of some variable.}

\item{colLink}{Default color of links.}

\item{sizeLink}{Default line width of links.}

\item{maxSizeLink}{Maximal line width of a link when it represents the value of some variable.}
}
\value{
A list with the values of the different graphical parameters.
}
\description{
This function gets and sets options that control some graphical aspects 
of maps created with \code{\link{plotMap}}
}
\examples{
\dontrun{
params <- plotMapOptions(colArea = "red", colLink = "orange")
plotMap(mydata, mylayout, options = params)
}

}

