% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotThermalGeneration.R
\name{plotThermalGeneration}
\alias{plotThermalGeneration}
\title{Plot for Thermal Capacities}
\usage{
plotThermalGeneration(data, area = "all", main = "Thermal generation")
}
\arguments{
\item{data}{data.table of Thermal capacities}

\item{area}{areas to select, default all}

\item{main}{title}
}
\description{
Plot for Thermal Capacities
}
\examples{
\dontrun{
opts <- setSimulationPath(getwd())
plotThermalGeneration( thermalGeneration(opts))
}

}
