% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_thermal_group_capacities.R
\name{plotThermalGroupCapacities}
\alias{plotThermalGroupCapacities}
\title{Plot for Thermal Group Capacities}
\usage{
plotThermalGroupCapacities(data, area = "all",
  main = "Thermal group capacities")
}
\arguments{
\item{data}{data.table of Thermal Group capacities}

\item{area}{areas to select, default all}

\item{main}{title}
}
\description{
Plot for Thermal Group Capacities
}
\examples{
\dontrun{
opts <- setSimulationPath(getwd())
plotThermalGroupCapacities( thermalGroupCapacities(opts))
}

}
