% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productionStack.R
\name{print.productionStack}
\alias{print.productionStack}
\title{Visualize the object generated by function productionStack}
\usage{
\method{print}{productionStack}(x, ...)
}
\description{
Visualize the object generated by function productionStack
}

