% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodStackAliases.R
\name{prodStackAliases}
\alias{prodStackAliases}
\title{Aliases of production stacks}
\usage{
prodStackAliases()
}
\value{
Print aliases.
}
\description{
This function shows aliases for \code{\link{prodStack}}. These aliases can be used
in the \code{\link{prodStack}} function in the \code{variables} argument
}
\examples{
prodStackAliases()

}

