% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productionStackAliases.R
\name{productionStackAliases}
\alias{productionStackAliases}
\title{show aliases for productionStack}
\usage{
productionStackAliases()
}
\value{
Print aliases.
}
\description{
This function shows aliases for productionStack. These aliases can be used
in the \code{\link{productionStack}} function in the \code{variables} argument
}
\examples{
productionStackAliases()

}

