% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productionStack.R
\name{productionStackLegend}
\alias{productionStackLegend}
\title{Plot an interactive legend for a production stack plot}
\usage{
productionStackLegend(variables = "eco2mix", colors = NULL, lines = NULL,
  lineColors = NULL, legendItemsPerRow = 5, legendId = "")
}
\arguments{
\item{variables}{Either a character string containing an alias or a named list of expressions 
created with \code{\link[base]{alist}}. The name of each element is the name
of the variable to draw in the stacked graph. The element itself is an
expression explaining how to compute the variable (see examples).}

\item{colors}{Vector of colors with same length as parameter \code{variables}. If 
\code{variables} is an alias, then this argument should be \code{NULL} in 
order to use default colors.}

\item{lines}{A named list of expressions created with \code{\link[base]{alist}}
indicating how to compute the curves to display on top of the stacked graph.
It should be \code{NULL} if there is no curve to trace or if parameter
\code{variables} is an alias.}

\item{lineColors}{Vector of colors with same length as parameter \code{lines}. This argument
should be \code{NULL} if there is no curve to trace or if parameter
\code{variables} is an alias.}

\item{legendItemsPerRow}{Number of elements to put in each row of the legend.}

\item{legendId}{Id of the legend linked to the graph. This argument is 
usefull when one wants to create a shared legend with
\code{\link{productionStackLegend}}}
}
\description{
This function create an nice looking legend that displays values when the user
hovers a production stack created with \code{\link{productionStack}}. By 
default \code{\link{productionStack}} already outputs a legend. This function
is mostly usefull to share legend between two or more production stacks.
}
\details{
This function can be used to create a legend shared by multiple production 
stacks in a Shiny application or an interactive document created with Rmarkdown.
For instance, let assume one wants to display four productions stacks in a 2x2
layout and have a unique legend below them in a Rmarkdown document. To do so,
one can use the following chunck code:

\preformatted{
```{R, echo = FALSE}
library(shiny)

fillCol(height = "600px", flex = c(1, 1, NA),
  fillRow(
    productionStack(mydata, areas = "fr", 
                    main = "Production stack in France", unit = "GWh", 
                    legend = FALSE, legendId = 1, height = "100\%"),
    productionStack(mydata, areas = "de", 
                    main = "Production stack in Germany", unit = "GWh", 
                    legend = FALSE, legendId = 1, height = "100\%"),
  ),
  fillRow(
    productionStack(mydata, areas = "es", 
                    main = "Production stack in Spain", unit = "GWh", 
                    legend = FALSE, legendId = 1, height = "100\%"),
    productionStack(mydata, areas = "be", 
                    main = "Production stack in Belgium", unit = "GWh", 
                    legend = FALSE, legendId = 1, height = "100\%"),
  ),
  productionStackLegend(legendId = 1)
)
```
}
}

