% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runAppAntaresViz}
\alias{runAppAntaresViz}
\title{Run app antaresViz}
\usage{
runAppAntaresViz(browser = FALSE)
}
\arguments{
\item{browser}{\code{boolean} run app in browser}
}
\value{
an App Shiny.
}
\description{
\code{runAppAntaresViz} run antaresViz App.
}
