% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInteractivity.R
\name{setInteractivity}
\alias{setInteractivity}
\alias{getInteractivity}
\title{Get and set interactivity mode}
\usage{
setInteractivity(interactive = "auto")

getInteractivity()
}
\arguments{
\item{interactive}{Should plot functions generate a UI that lets users to 
interactively modify input data and graphical parameters of a chart? It should
be TRUE or FALSE. The default behavior is to set it to TRUE if the R session
is interactive and to FALSE otherwise (for instance in Rmarkdown document).}
}
\value{
\code{getInteractivity} returns a boolean indicating the interactivity mode of
plot functions. \code{setInteractivity} is only used for its side effects.
}
\description{
\code{setInteractivity} globally sets the interactivity mode of plot 
functions. This is useful to avoid repeating \code{interactive = FALSE} or
\code{interactive = TRUE} in each function. \code{getInteractivity} gets the
interactivity mode.
}
