% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_map.R
\name{stackMap}
\alias{stackMap}
\title{plot stack and map}
\usage{
stackMap(x, mapLayout)
}
\arguments{
\item{x}{\code{antaresDataList} antaresDataList contian areas ans links.}

\item{mapLayout}{Object created with function \code{\link{mapLayout}}}
}
\description{
plot stack and map
}
\examples{
\dontrun{
mydata <- readAntares(areas = "all", links = "all")

layout <- readLayout()
ml <- mapLayout(layout = layout)

stackMap(x = mydata, mapLayout = ml)
}

}
