% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapUtils.R
\name{updateCircleMarkers}
\alias{updateCircleMarkers}
\title{Update the style of circle markers}
\usage{
updateCircleMarkers(map, layerId, radius = NULL, stroke = NULL,
  color = NULL, weight = NULL, opacity = NULL, fill = NULL,
  fillColor = NULL, fillOpacity = NULL)
}
\arguments{
\item{map}{A map object created with \code{\link[leaflet]{leaflet()}}}

\item{layerId}{The layer Id}

\item{radius}{Numeric vector of radii for the circles.}

\item{stroke}{Wheter to draw the border of the circles}

\item{color}{Color of the border of the circles}

\item{weight}{Line width of the border of the circles}

\item{opacity}{Opacity of the border of the circles}

\item{fill}{Wheter to fill the circles with colors}

\item{fillColor}{Fill color}

\item{fillOpacity}{Fill opacity}
}
\value{
the new map object
}
\description{
Function to update the visual properties of circle markers already added to a 
map. The main utility of this function appears when used with function 
\code{\link[leaflet]{leafletProxy}}.
}
\examples{
require(leaflet)

leaflet() \%>\% 
  addCircleMarkers(c(0, 0), c(0, 10), layerId = c("a", "b")) \%>\% 
  updateCircleMarkers(c("a"), color = "red")

\dontrun{  
# updateCircleMarkers becomes really usefull inside the server function of
# a shiny application.

server <- function(input, output, session) {
  output$map <- renderLeaflet({
    leaflet() \%>\% addCircleMarkers(c(0, 0), c(0, 10), layerId = c("a", "b"))
  })
  
  map <- leafletProxy("map", session)
  
  observe({
    updateCircleMarkers(map, layerId = c("a", "b"), color = input$color)
  })
  
  ...
}

}
}

