% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limitSizeGraph.R
\name{limitSizeGraph}
\alias{limitSizeGraph}
\title{Use to change limit size of graph (in Mb)}
\usage{
limitSizeGraph(size)
}
\arguments{
\item{size}{\code{numeric} widget size autorized in modules (default 200)}
}
\description{
Use to change limit size of graph (in Mb)
}
\examples{
\dontrun{
limitSizeGraph(500)
}

}
