% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layout.R
\name{mapLayout_no_interactive}
\alias{mapLayout_no_interactive}
\title{Build a `mapLayout` object with gesojson file}
\usage{
mapLayout_no_interactive(path_geojson_file, opts = simOptions())
}
\arguments{
\item{path_geojson_file}{`character` path of geojson file}

\item{opts}{list of simulation parameters returned by 
the function [antaresRead::setSimulationPath()]}
}
\description{
this function creates a 'mapLayout' object from a study and an external 
'geojson' file. 
The 'geojson' file must contain zones compatible with the study. 
This function should be used only once per study. 
The result should then be saved in an external file and be reused.
}
