% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMapOptions.R
\name{plotMapOptions}
\alias{colorScaleOptions}
\alias{plotMapOptions}
\title{Graphical options for plotMap}
\usage{
plotMapOptions(areaDefaultCol = "#CCCCCC", areaDefaultSize = 15,
  areaMaxSize = 15, areaChartColors = NULL, areaChartUniqueScale = FALSE,
  areaColorScaleOpts = colorScaleOptions(), linkDefaultCol = "#CCCCCC",
  linkDefaultSize = 3, linkMaxSize = 10,
  linkColorScaleOpts = colorScaleOptions())

colorScaleOptions(n = 5, domain = NULL, negCol = "#FF0000",
  zeroCol = "#FFFFFF", posCol = "#0000FF", naCol = "#EEEEEE",
  zeroTol = NULL)
}
\arguments{
\item{areaDefaultCol}{default color of areas.}

\item{areaDefaultSize}{default size of areas.}

\item{areaMaxSize}{maximal size of an area when it represents the value of some variable.}

\item{areaChartColors}{Vector of colors to use in polar area charts and bar charts}

\item{areaChartUniqueScale}{If the map contains polar or bar charts, should the different variables 
represented use the same scale or should each variable have its own scale ?
This parameter should be TRUE only if the variables have the same unit and 
are comparable : for instance production variables.}

\item{areaColorScaleOpts}{List of options used to construct a continuous color scale. This list should
be generated with function \code{colorScaleOptions}.}

\item{linkDefaultCol}{Default color of links.}

\item{linkDefaultSize}{Default line width of links.}

\item{linkMaxSize}{Maximal line width of a link when it represents the value of some variable.}

\item{linkColorScaleOpts}{List of options used to construct a continuous color scale. This list should
be generated with function \code{colorScaleOptions}.}

\item{n}{Approximate number of colors to use. The function tries to cut the data 
nicely, so the real number of colors used may vary.}

\item{domain}{Range of the data, ie. the range of possible values. If \code{NULL}, the
the range of the data is used}

\item{negCol}{color of the extreme negative value.}

\item{zeroCol}{color of the 0 value.}

\item{posCol}{Color of the extreme positive value.}

\item{naCol}{Color for missing values}

\item{zeroTol}{All values in the interval \code{\[-zeroTol, +zeroTol\]} are mapped to the
\code{zeroCol} color. If \code{NULL}, the function tries to pick a nice
value that is approximately equal to 1\% of the maximal value.}
}
\value{
A list with the values of the different graphical parameters.
}
\description{
These functions get and set options that control some graphical aspects 
of maps created with \code{\link{plotMap}}.
}
\examples{
\dontrun{
params <- plotMapOptions(areaDefaultCol = "red", linkDefaultCol = "orange")
plotMap(mydata, mylayout, options = params)
}

}

