% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotThermalGroupCapacities.R
\name{plotThermalGeneration}
\alias{plotThermalGeneration}
\title{Plot for Thermal Group Capacities}
\usage{
plotThermalGeneration(data, area = "all", main = "Thermal group capacities")
}
\arguments{
\item{data}{data.table of Thermal Group capacities}

\item{area}{areas to select, default all}

\item{main}{title}
}
\description{
Plot for Thermal Group Capacities
}
\examples{
\dontrun{
opts <- setSimulationPath(getwd())
plotThermalGroupCapacities( thermalGroupCapacities(opts))
}

}
