% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_XY.R
\name{plotXY}
\alias{plotXY}
\title{Plot density between X et Y with ggplot2 and plotly}
\usage{
plotXY(
  data,
  x,
  y,
  precision = 30,
  sizeOnCount = FALSE,
  outLine = TRUE,
  transform = NULL
)
}
\arguments{
\item{data}{\code{data.frame} can be antaresData object}

\item{x}{\code{character}, x variable}

\item{y}{\code{character}, y variable}

\item{precision}{Deprecated.}

\item{sizeOnCount}{Deprecated.}

\item{outLine}{Deprecated.}

\item{transform}{Deprecated.}
}
\description{
Plot density between X et Y with ggplot2 and plotly
}
\examples{
\dontrun{

setSimulationPath("myStudy")
myData <- readAntares()

plotXY(myData, "NODU", "LOAD", precision = 50,
         sizeOnCount = FALSE)
         
myData <- readAntares(areas = "all", links = "all")    
myData <- mergeAllAntaresData(myData)  
plotXY(myData, "OP. COST_max_b", "OP. COST_max_c", precision = 50,
         sizeOnCount = FALSE)
         
         
}
   
}
